Promote @types/express from devDeps to deps
That is, from devDependencies to regular dependencies.

This is needed because we export an interface (serverRoutes()) that
depends on those types.  Without this change, the monorepo's
`test-publish`, running with an asset-swapper modified to depend on this
package, gave:

node_modules/@0x/quote-server/lib/src/router.d.ts:2:70 - error TS2307: Cannot find module 'express-serve-static-core'.

2 export declare const serverRoutes: (quoteStrategy: Quoter) => import("express-serve-static-core").Router;
                                                                       ~~~~~~~~~~~~~~~~~~~~~~~~~~~

("express-serve-static-core" is a direct dependency of @types/express.)